Switch back to using Docker
The systemd service was causing odd hangs with the WiFi network and requiring a reboot of the Pi
I wanted to figure out why Docker wasn't working regardless. Turns out the go-rpio library import in hardware_sensor.go had the version package removed.  Added the 'V4' back and was able to control the gpio again.